align types

+++ hs_detection/detect/Types.h
@@ -1,24 +1,23 @@
 #ifndef TYPES_H
 #define TYPES_H
 
-// #include <cstdint>
+#include <cstdint>
 #include <limits>
 
-// TODO: replace with int32_t?
 #define MAX_FRAME std::numeric_limits<IntFrame>::max()
 
 namespace HSDetection
 {
-    typedef int IntFrame;
-    typedef int IntChannel;
-    typedef short IntVolt;
-    typedef float FloatGeom; // TODO: float32? maybe no need
+    typedef int32_t IntFrame;
+    typedef int32_t IntChannel;
+    typedef int16_t IntVolt;
+    typedef float FloatGeom; // correspond to np.single and cython.float
 
-    typedef long IntFxC;
-    typedef int IntFxV;
-    typedef int IntCxV;
-    typedef int IntFCV;
-    typedef int IntResult; // TODO: ??? or FxC
+    typedef int64_t IntFxC;
+    typedef int32_t IntFxV;
+    typedef int32_t IntCxV;
+    typedef int64_t IntFCV;
+    typedef int32_t IntResult; // TODO: ??? or FxC
 
 } // namespace HSDetection
